Rewrite Image component and its utility methods in TS.
