--- src/Image/index.js
@@ -1,2 +1,0 @@
-import Image from './Image';
-export default Image;
