+++ src/Image/index.ts
@@ -1,0 +1,5 @@
+import Image from './Image';
+import getUrl, { getBackgroundUrl } from './getUrl';
+
+export default Image;
+export { getUrl, getBackgroundUrl };
